fix: prevent  extra interceptors in strict mode
